move functionality out of speed and utils_graph modules into better homes
