+++ docs/source/user-reference.rst
@@ -13,6 +13,12 @@ osmnx.bearing module
 .. automodule:: osmnx.bearing
     :members:
 
+osmnx.convert module
+--------------------
+
+.. automodule:: osmnx.convert
+    :members:
+
 osmnx.distance module
 ---------------------
 
@@ -79,12 +85,6 @@ osmnx.simplification module
 .. automodule:: osmnx.simplification
     :members:
 
-osmnx.speed module
-------------------
-
-.. automodule:: osmnx.speed
-    :members:
-
 osmnx.stats module
 ------------------
 
@@ -107,10 +107,4 @@ osmnx.utils_geo module
 ----------------------
 
 .. automodule:: osmnx.utils_geo
-    :members:
-
-osmnx.utils_graph module
-------------------------
-
-.. automodule:: osmnx.utils_graph
     :members:
