+++ osmnx/simplification.py
@@ -13,10 +13,10 @@ from shapely.geometry import MultiPolygon
 from shapely.geometry import Point
 from shapely.geometry import Polygon
 
+from . import convert
 from . import settings
 from . import stats
 from . import utils
-from . import utils_graph
 from ._errors import GraphSimplificationError
 
 if TYPE_CHECKING:
@@ -544,7 +544,7 @@ def _merge_nodes_geometric(G: nx.MultiDiGraph, tolerance: float) -> gpd.GeoSerie
         The merged overlapping polygons of the buffered nodes.
     """
     # buffer nodes GeoSeries then get unary union to merge overlaps
-    merged = utils_graph.graph_to_gdfs(G, edges=False)["geometry"].buffer(tolerance).unary_union
+    merged = convert.graph_to_gdfs(G, edges=False)["geometry"].buffer(tolerance).unary_union
 
     # if only a single node results, make it iterable to convert to GeoSeries
     merged = MultiPolygon([merged]) if isinstance(merged, Polygon) else merged
@@ -603,7 +603,7 @@ def _consolidate_intersections_rebuild_graph(  # noqa: C901,PLR0912,PLR0915
     # attach each node to its cluster of merged nodes. first get the original
     # graph's node points then spatial join to give each node the label of
     # cluster it's within. make cluster labels type string.
-    node_points = utils_graph.graph_to_gdfs(G, edges=False)
+    node_points = convert.graph_to_gdfs(G, edges=False)
     cols = set(node_points.columns).intersection(["geometry", *settings.useful_tags_node])
     node_points = node_points[list(cols)]
     gdf = gpd.sjoin(node_points, node_clusters, how="left", predicate="within")
@@ -680,7 +680,7 @@ def _consolidate_intersections_rebuild_graph(  # noqa: C901,PLR0912,PLR0915
 
     # STEP 6
     # create new edge from cluster to cluster for each edge in original graph
-    gdf_edges = utils_graph.graph_to_gdfs(G, nodes=False)
+    gdf_edges = convert.graph_to_gdfs(G, nodes=False)
     for u, v, k, data in G.edges(keys=True, data=True):
         u2 = gdf.loc[u, "cluster"]
         v2 = gdf.loc[v, "cluster"]
